Backend: Check if dataset is being used by a Widget before deleting (return 409 error if so) (#198)
* Backend: Check if dataset is being used by a Widget before deleting (return 409 error if so)

* Backend: Check if dataset is being used by a Widget before deleting (return 409 error if so)

* Add FilterExpression and pagination

+++ backend/src/lib/controllers/ingestapi-ctrl.ts
@@ -87,6 +87,12 @@ async function deleteDataset(req: Request, res: Response) {
   const { id } = req.params;
 
   const repo = DatasetRepository.getInstance();
+  const widgetCount = await repo.getWidgetCount(id);
+
+  if (widgetCount > 0) {
+    return res.status(409).send("Dataset has widgets associated to it");
+  }
+
   await repo.deleteDataset(id);
   logger.info("Dataset deleted %s", id);
